DRILL-1942 improve direct memory usage tracking Implemented a new allocator (BaseAllocator, ChildAllocator, RootAllocator) that is much stricter than TopLevelAllocator. New tests for the new allocators in TestBaseAllocator.
The new allocator flattens the use of buffers in slices,
so that a DrillBuf is always referring to an UnsafeDirectLittleEndian, and
never to another DrillBuf; this simplifies figuring out what's going on when
something goes wrong. The new allocator (BaseAllocator, and its derived
classes RootAllocator and ChildAllocator) treats sharing buffers in the same
way as slicing buffers, and sharing and ownership transfer of buffers and
their slices both work. After exhausting its initial allotment, the new
allocator asks its parent for larger chunks instead of going to its parent
for each subsequent request.

- fixed a number of leaks
- converted many classes to AutoCloseable (often renaming methods like
cleanup() to close()) in order to have the compiler detect resource leaks.
- made pre allocation reservations AutoCloseable
- introduced the use of a factory for allocator instantiation so that
  one of multiple allocators can be chosen at run-time