DRILL-1950: Parquet rowgroup level filter pushdown.
Implement Parquet rowgroup level filter pushdown. The filter pushdown is performed in
query planning time, in Drill physical planning phase.

Only a local filter, which refers to columns in a single table, is qualified for filter pushdown.

A filter may be elligible if it is a simple compare filter, or a compound "and/or" filter consists of
simple compare filter. Data types allowed in compare filter are int, bigint, float, double, date,
timestamp, time. Compare operators are =, !=, <, <=, >, >=. Operands have to be a column of the above
data types, or an explicit cast or implicit cast function, or a constant expressions.