CAST function index support (#61)
* IndexDescriptor uses more generic LogicalExpression for indexedField, thus we could index more generic expression(e.g. cast functions).

We also need to check if an expression in filter condition is 'indexable' or not, which means
only the expressions which were operands in comparison operators that involve at least one column can be indexed.
e.g. for this condition "concat(column2, 'string1')='aaa'", even column2 is indexed, this filter condition could not utilize
the index, but if "concat(column2, 'string1')" is indexed, then this condition can utilize that index.

1, for cast-to-varchar index, if the in query a cast-to-varchar has shorter length, we convert the equality comparison to LIKE
2, for covering plan, use global knowledge(full original logical plan cached) to decide if we should create covering plan, because when index is a function on schemapath, we could not decide if the schemapath was refered somewhere out of the indexx rule's scope(filter-project-scan).

LIMIT: While indexable expression could be more generic other than plain field, we are limiting it to only function(and only CAST for V1). the limit was enforced for the simplicity in defining indexes.

Remove Covering Plan for now

* Address comments

* comment remove mapr specific

* Address comments